\name{npcure-package}

\alias{npcure-package}

\alias{npcure}

\docType{package}

\title{\packageTitle{npcure}}

\description{\packageDescription{npcure}}

\details{\packageIndices{npcure}}

\author{Ignacio López-de-Ullibarri [aut, cre],
  Ana López-Cheda [aut],
  Maria Amalia Jácome [aut]
}

\references{López-Cheda, A., Cao, R., Jácome, M. A., Van Keilegom,
  I. (2017). Nonparametric incidence estimation and bootstrap bandwidth
  selection in mixture cure models. Computational Statistics & Data
  Analysis 105: 144–165. \url{https://doi.org/10.1016/j.csda.2016.08.002}.

  López-Cheda, A., Jácome, M. A., Cao, R. (2017). Nonparametric latency
  estimation for mixture cure models. Test, 26:
  353–376. \url{https://doi.org/10.1007/s11749-016-0515-1}.
}

\keyword{package}

\concept{Nonparametric Mixture Cure Model}
